+++ src/App.js
@@ -1,35 +1,15 @@
 import Home from "./routes/Home";
 import Detail from "./routes/Detail";
 import React, { useEffect, useState } from "react";
-import { BrowserRouter, Routes, Route } from "react-router-dom";
-import styled from "styled-components";
-
-import { useSelector, useDispatch } from "react-redux";
+import { Routes, Route } from "react-router-dom";
 
 function App() {
-  const isDarkMode = useSelector((state) => state.isDarkMode);
-  const dispatch = useDispatch();
-
-  const handleDarkMode = () => {
-    dispatch({ type: "DarkMode" });
-  };
-
   return (
-    <AppContainer isDarkMode={isDarkMode}>
-      <button onClick={handleDarkMode}>
-        {isDarkMode ? "WhiteMode" : "DarkMode"}
-      </button>
-      <Routes>
-        <Route path="/" element={<Home />} />
-        <Route path="/detail/:id" element={<Detail />} />
-      </Routes>
-    </AppContainer>
+    <Routes>
+      <Route path="/" element={<Home />} />
+      <Route path="/detail/:id" element={<Detail />} />
+    </Routes>
   );
 }
 
 export default App;
-
-const AppContainer = styled.div`
-  background-color: ${(props) => (props.isDarkMode ? "#000" : "#fff")};
-  color: ${(props) => (props.isDarkMode ? "#fff" : "#000")};
-`;
