--- src/components/Button.js
@@ -1,18 +1,0 @@
-// import { useSelector, useDispatch } from "react-redux";
-
-// const Button = () => {
-//   const isDarkMode = useSelector((state) => state.isDarkMode);
-//   const dispatch = useDispatch();
-
-//   const handleDarkMode = () => {
-//     dispatch("DarkMode");
-//   };
-
-//   return (
-//     <div>
-//       <buuton>{isDarkMode ? "DarkMode" : "WhiteMode"}</buuton>
-//     </div>
-//   );
-// };
-
-// export default Button;
